Finish adding extra points such as:
Defensive logic to prevent bugs when pressing
equals button prematurely. The event listener only
responds if the secondOperand exists.

Keyboard supoprt works for every key: arithmetic
functions, number functions, clearing and deleting.

For HTML, the ID of the decimal was changed to '.'
in order for the code recognize it based on the
e.key

I tried with e.code but since there are two keys
that write the period in the keyboard, it overcomplicated
the conditional.