Fix operator bug & refactor appendOperand
1. Operator buttons evaluate an expression with the firstOperand
whenever the same operator is pressed twice, e.g. 8+ -> + -> 16

but the issue was that it ignored if a secondOperand had been passed:
8+2 -> + -> 16.

I implemented checking if the secondOperand exists to just not doing
anything when pressing any operator if the two operands exist.

2. Previously, I had two, almost identical, blocks of if/else if
in the appendOperand function. I have simplified the function by
declaring a targetOperandValue that determines which operand should be
appended by checking if operator exists.